+++ migrations/00026_add_game_notes.sql
@@ -1,0 +1,9 @@
+-- +goose Up
+-- +goose StatementBegin
+ALTER TABLE game_installs ADD COLUMN notes TEXT;
+-- +goose StatementEnd
+
+-- +goose Down
+-- +goose StatementBegin
+ALTER TABLE game_installs DROP COLUMN notes;
+-- +goose StatementEnd